+++ netcode/shared/packets/tcp/tcp_packet_header.hpp
@@ -6,5 +6,5 @@
 /* Header MUST ALWAYS be the first field of the packet for deserializing type. */
 struct __attribute__((packed)) TCPPacketHeader {
     TCPPacketType type;
-    uint8_t payloadSize;
+    uint16_t payloadSize;
 };
