+++ netcode/shared/packets/udp/udp_packet.hpp
@@ -39,7 +39,7 @@ public:
     }
 
     template<typename T>
-    static T create(const UDPPacketType type, const uint32_t id, const char *payload, const uint8_t payloadSize) {
+    static T create(const UDPPacketType type, const uint32_t id, const char *payload, const uint16_t payloadSize) {
         if (payloadSize > MAX_UDP_PAYLOAD_SIZE) {
             throw std::length_error("Payload size exceeds MAX_UDP_PAYLOAD_SIZE");
         }
