+++ netcode/shared/packets/udp/udp_packet_header.hpp
@@ -6,6 +6,6 @@
 /* Header MUST ALWAYS be the first field of the packet for deserializing type. */
 struct __attribute__((packed)) UDPPacketHeader {
     UDPPacketType type;
-    uint8_t payloadSize;
+    uint16_t payloadSize;
     uint32_t id;
 };
